Adapted for compliance.
